Change track_log to local_track_list
I switched the use of the quicksaver track log list to the a local track
list set by getting the references to the track lists from controller.

If I realize later that it's better to keep them separate, especially
because the log was a list that maintained order, then I can switch back
by referring to this commit.