Add functions for OS errors, unexpected errors, and critical errors
I added a function for indicating a critical error, which is used any
time there's a critical error where the Pico has to be shut down.
The error's dedicated notifier will be triggered first to indicate the
specific problem, and then the critical error will be triggered to
indicate that the program will be shutting down.

I also added a function for triggering OS errors, and unexpected OS
errors, as well as renamed trigger_led to flash_led.