Pass GPIO pin numbers to listener and notifier using a dict
Rather than defining parameters for each GPIO device (LEDs and buttons)
and passing the pin numbers for each one, I refactored it to recieve a
dict of the GPIO pins that can be unpacked by each constructor.
This works great because I can just grab the GPIO pins dict from the
config file as-is and just pass it to the constructors of the notifier
and listener. It's also more concise than having all the parameters.