Add Button class with debouncing logic
We implemented button interrupts in the last commit, but this time we
refactored it to add debouncing logic. In doing that, we created a new
Button class that encapsulates all the button pin code and debouncing.
We also removed the start and stop listener methods because they are not
needed.